Unsigned int instead of int

+++ 0x14-bit_manipulation/4-clear_bit.c
@@ -7,7 +7,7 @@
  */
 int clear_bit(unsigned long int *n, unsigned int index)
 {
-	int mask = 1 << index;
+	unsigned long int mask = 1 << index;
 
 	if (sizeof(n) * 8)
 	{
